+++ core/src/test/java/de/bitbrain/braingdx/world/GameWorldTest.java
@@ -1,12 +1,10 @@
 package de.bitbrain.braingdx.world;
 
-import com.badlogic.gdx.graphics.OrthographicCamera;
 import de.bitbrain.braingdx.util.GdxUtils;
 import de.bitbrain.braingdx.util.Mutator;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
-import org.mockito.Mock;
 import org.mockito.runners.MockitoJUnitRunner;
 
 import java.util.Comparator;
@@ -17,15 +15,12 @@ import static org.assertj.core.api.Assertions.assertThat;
 @RunWith(MockitoJUnitRunner.class)
 public class GameWorldTest {
 
-   @Mock
-   private OrthographicCamera camera;
-
    private GameWorld world;
 
    @Before
    public void beforeTest() {
       GdxUtils.mockApplicationContext();
-      world = new GameWorld(camera);
+      world = new GameWorld();
    }
 
    @Test
@@ -131,5 +126,31 @@ public class GameWorldTest {
          }
       });
       assertThat(sorted).containsExactly(object1, object3, object2);
+   }
+
+   @Test
+   public void testAddToGroup() {
+      GameObject object1 = world.addObject("asdf");
+      GameObject object2 = world.addObject("asdf");
+      assertThat(world.getGroup("asdf")).containsExactly(object1, object2);
+   }
+
+   @Test
+   public void testRemoveFromGroup() {
+      GameObject object1 = world.addObject("asdf");
+      GameObject object2 = world.addObject("asdf");
+      world.remove(object1);
+      assertThat(world.getGroup("asdf")).containsExactly(object2);
+   }
+
+   @Test
+   public void testClearGroup() {
+      world.addObject("asdf");
+      world.addObject("asdf");
+      GameObject object3 = world.addObject("asdf2");
+      GameObject object4 = world.addObject("asdf2");
+      world.clearGroup("asdf");
+      assertThat(world.getGroup("asdf")).isEmpty();
+      assertThat(world.getObjects()).containsExactlyInAnyOrder(object3, object4);
    }
 }